requirements.txt
asttokens==2.4.1
attrs==23.1.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
cached-property==1.5.2
certifi==2023.11.17
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.0
comtypes==1.2.0
contourpy==1.1.1
convertdate==2.4.0
cryptography==41.0.7
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.4.2
docutils==0.20.1
element-manager==1.0.4
et-xmlfile==1.1.0
exceptiongroup==1.2.0
exchangelib==4.9.0
executing==2.0.1
Faker==20.1.0
fire==0.4.0
fonttools==4.47.0
fpdf2==2.7.7
furl==2.1.3
graphviz==0.13.2
greenlet==3.0.3
h11==0.14.0
hijri-converter==2.3.1
holidays==0.21.13
htmldocx==0.0.6
idna==3.6
importlib-metadata==7.0.0
importlib-resources==6.1.1
ipdb==0.13.13
ipykernel==6.27.1
ipython==8.12.3
isodate==0.6.1
java-access-bridge-wrapper==1.1.0
jedi==0.19.1
jsonpath-ng==1.6.0
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter_client==8.6.0
jupyter_core==5.5.0
kiwisolver==1.4.5
korean-lunar-calendar==0.3.1
lxml==4.9.4
matplotlib==3.7.4
matplotlib-inline==0.1.6
more-itertools==10.1.0
mouse==0.7.1
MouseInfo==0.1.3
mss==6.1.0
nest-asyncio==1.5.8
netsuitesdk==1.24.0
notifiers==1.3.3
numpy==1.24.4
O365==2.0.26
oauthlib==3.2.2
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
openpyxl==3.1.2
orderedmultidict==1.0.1
outcome==1.3.0.post0
packaging==23.2
pandas==2.0.3
parso==0.8.3
pdfminer.six==20221105
pendulum==2.1.2
pickleshare==0.7.5
Pillow==10.1.0
pkgutil_resolve_name==1.3.10
platformdirs==4.1.0
ply==3.11
prompt-toolkit==3.0.41
psutil==5.9.6
psycopg2==2.9.9
pure-eval==0.2.2
PyAutoGUI==0.9.54
pycparser==2.21
PyGetWindow==0.0.9
Pygments==2.17.2
PyJWT==2.8.0
PyMeeus==0.5.12
PyMsgBox==1.0.9
pynput==1.7.6
pynput-robocorp-fork==5.0.0
pyotp==2.9.0
pyparsing==3.1.1
pypdf==3.17.3
pyperclip==1.8.2
pypiwin32==223
PyRect==0.2.0
PyScreeze==0.1.30
PySide2==5.15.2.1
PySimpleGUI==4.60.5
PySocks==1.7.1
pyspnego==0.10.2
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
pytweening==1.0.7
pytz==2023.3.post1
pytz-deprecation-shim==0.1.0.post0
pytzdata==2020.1
pywin32==306
pywinauto==0.6.8
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.32.0
requests==2.31.0
requests-file==1.5.1
requests-ntlm==1.2.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
robocorp-storage==1.0.2
robotframework==6.1.1
robotframework-pythonlibcore==4.3.0
robotframework-requests==0.9.6
robotframework-sapguilibrary==1.1
robotframework-seleniumlibrary==6.2.0
robotframework-seleniumtestability==2.1.0
rpaframework==28.0.0
rpaframework-core==11.3.0
rpaframework-pdf==7.3.1
rpaframework-recognition==5.2.3
rpaframework-windows==7.5.0
rpapy==1.0.2
rpds-py==0.15.2
selenium==4.15.2
shiboken2==5.15.2.1
simple-salesforce==1.12.5
six==1.16.0
smartsheet-python-sdk==3.0.2
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.25
sspilib==0.1.0
stack-data==0.6.3
stringcase==1.2.0
summarytools==0.2.3
tenacity==8.2.3
termcolor==2.4.0
tomli==2.0.1
tornado==6.4
traitlets==5.14.0
trio==0.23.1
trio-websocket==0.11.1
tweepy==3.10.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==4.3.1
uiautomation==2.0.18
urllib3==2.1.0
validators==0.20.0
wcwidth==0.2.12
webdriver-manager==4.0.1
win10toast==0.9
wrapt==1.16.0
wsproto==1.2.0
xlrd==2.0.1
xlutils==2.0.0
xlwt==1.3.0
zeep==4.2.1
zipp==3.17.0
